Add noopener along with _target blank for security
This is for security reasons. Without it if a user clicks an external
link the target page will have full access to the `window` object!

To fix this we add `rel="noopener noreferrer"` along with
`_target="blank"`. `noreferrer` is for older browsers that does not
support `noopener`.

For more information see https://mathiasbynens.github.io/rel-noopener/